refactor: move BrandChallengePromptsTooltip component and update LazyBrandChallengePromptsTooltip import to avoid circular reference
